Refactor ProgressView with modular architecture and improved state management
