+++ src/progressView/ProgressStateManager.old.ts
